Try to include remainder of tests in last bundle
When packing, if we exceed optimalBundleTime but are on the last bundle
it's usually faster to just add the remainders to the last bundle
than have to run a bundle of just a few tests. The heuristic
for now is if the remaining tests can be done in less than
optimalBundleTime tack them on, but this could be refined.
For example to only add the remainders if it's a small fraction
of optimalBundleTime.